+++ backend/src/test/java/com/allog/dallog/presentation/ControllerTest.java
@@ -3,14 +3,19 @@ package com.allog.dallog.presentation;
 import com.allog.dallog.common.config.ExternalApiConfig;
 import com.allog.dallog.domain.auth.application.AuthService;
 import com.allog.dallog.domain.auth.application.OAuthUri;
+import com.allog.dallog.domain.auth.presentation.AuthController;
 import com.allog.dallog.domain.category.application.CategoryService;
+import com.allog.dallog.domain.category.presentaion.CategoryController;
 import com.allog.dallog.domain.categoryrole.application.CategoryRoleService;
 import com.allog.dallog.domain.externalcalendar.application.ExternalCalendarService;
+import com.allog.dallog.domain.externalcalendar.presentation.ExternalCalendarController;
 import com.allog.dallog.domain.member.application.MemberService;
+import com.allog.dallog.domain.member.presentation.MemberController;
 import com.allog.dallog.domain.schedule.application.CheckedSchedulesFinder;
 import com.allog.dallog.domain.schedule.application.ScheduleService;
+import com.allog.dallog.domain.schedule.presentation.ScheduleController;
 import com.allog.dallog.domain.subscription.application.SubscriptionService;
-import com.allog.dallog.presentation.auth.AuthController;
+import com.allog.dallog.domain.subscription.presentation.SubscriptionController;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
